Tweaked exposed overloads and refactored code.
